feat: flat bundles and size reductions
Use rollup to produce flat bundles of yup.
