+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0-beta.0",
   "description": "Dead simple Object schema validation",
   "main": "lib/index.js",
-  "module": "es/index.js",
+  "module": "lib/index.esm.js",
   "runkitExampleFilename": "./runkit-example.js",
   "scripts": {
     "test": "yarn lint && yarn test-all --runInBand",
@@ -15,13 +15,10 @@
     "precommit": "lint-staged",
     "toc": "doctoc README.md --github",
     "release": "rollout",
-    "build": "yarn 4c build && yarn toc",
+    "build:dts": "yarn tsc --emitDeclarationOnly -p . --outDir dts",
+    "build": "rm -rf dts && yarn build:dts && yarn rollup -c rollup.config.js  && yarn toc",
     "prepublishOnly": "yarn build"
   },
-  "files": [
-    "es",
-    "lib"
-  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/jquense/yup.git"
@@ -35,7 +32,8 @@
   },
   "homepage": "https://github.com/jquense/yup",
   "release": {
-    "conventionalCommits": true
+    "conventionalCommits": true,
+    "publishDir": "lib"
   },
   "prettier": {
     "singleQuote": true,
@@ -69,41 +67,41 @@
     "@4c/cli": "^3.0.1",
     "@4c/rollout": "^3.0.1",
     "@4c/tsconfig": "^0.4.0",
-    "@babel/cli": "^7.15.7",
-    "@babel/core": "^7.15.8",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@types/jest": "^27.0.3",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^27.2.5",
-    "babel-plugin-transform-rename-import": "^2.3.0",
-    "babel-preset-jason": "^6.3.0",
-    "doctoc": "^2.0.1",
-    "eslint": "^7.12.0",
-    "eslint-config-jason": "^8.1.1",
+    "@babel/cli": "^7.16.7",
+    "@babel/core": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-node-resolve": "^13.1.2",
+    "@types/jest": "^27.4.0",
+    "@typescript-eslint/eslint-plugin": "^5.8.1",
+    "@typescript-eslint/parser": "^5.8.1",
+    "babel-jest": "^27.4.5",
+    "babel-preset-env-modules": "^1.0.1",
+    "doctoc": "^2.1.0",
+    "dts-bundle-generator": "^6.3.0",
+    "eslint": "^8.6.0",
+    "eslint-config-jason": "^8.2.2",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jest": "^25.0.1",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-ts-expect": "^2.0.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jest": "^25.3.4",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-ts-expect": "^2.1.0",
     "eslint-plugin-typescript": "^0.14.0",
     "hookem": "^2.0.1",
-    "jest": "^27.2.5",
-    "lint-staged": "^11.2.3",
-    "prettier": "^2.4.1",
-    "rollup": "^2.58.0",
+    "jest": "^27.4.5",
+    "lint-staged": "^12.1.5",
+    "prettier": "^2.5.1",
+    "rollup": "^2.62.0",
     "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-dts": "^4.1.0",
     "rollup-plugin-filesize": "^9.1.1",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-size-snapshot": "^0.12.0",
     "synchronous-promise": "^2.0.15",
-    "typescript": "^4.4.3"
+    "typescript": "^4.5.4"
   },
   "dependencies": {
-    "nanoclone": "^0.2.1",
+    "nanoclone": "^1.0.0",
     "property-expr": "^2.0.4",
     "tiny-case": "^1.0.2",
     "toposort": "^2.0.2"
