+++ rollup.config.js
@@ -1,14 +1,16 @@
-const nodeResolve = require('rollup-plugin-node-resolve');
-const babel = require('rollup-plugin-babel');
-const filesize = require('rollup-plugin-filesize');
+import nodeResolve from '@rollup/plugin-node-resolve';
+import babel from '@rollup/plugin-babel';
+import dts from 'rollup-plugin-dts';
+import filesize from 'rollup-plugin-filesize';
 
 const base = {
-  input: './src/index.js',
+  input: './src/index.ts',
   plugins: [
-    nodeResolve(),
+    nodeResolve({ extensions: ['.js', '.ts'] }),
     babel({
-      babelrc: false,
-      presets: [['jason', { modules: false, runtime: false }]],
+      babelrc: true,
+      envName: 'esm',
+      extensions: ['.js', '.ts'],
     }),
   ],
   external: ['tiny-case', 'toposort', 'fn-name', 'property-expr'],
@@ -16,14 +18,19 @@ const base = {
 
 module.exports = [
   {
+    input: './dts/index.d.ts',
+    output: [{ file: 'lib/index.d.ts', format: 'es' }],
+    plugins: [dts()],
+  },
+  {
     ...base,
     output: [
       {
-        file: 'dist/yup.js',
+        file: 'lib/index.js',
         format: 'cjs',
       },
       {
-        file: 'dist/yup.esm.js',
+        file: 'lib/index.esm.js',
         format: 'es',
       },
     ],
