+++ src/setLocale.ts
@@ -1,12 +1,6 @@
-import locale from './locale';
+import locale, { LocaleObject } from './locale';
 
-type DeepPartial<T> = {
-  [P in keyof T]?: DeepPartial<T[P]>;
-};
-
-type Locale = DeepPartial<typeof locale>;
-
-export default function setLocale(custom: Locale) {
+export default function setLocale(custom: LocaleObject) {
   Object.keys(custom).forEach((type) => {
     // @ts-ignore
     Object.keys(custom[type]!).forEach((method) => {
