+++ src/util/isAbsent.ts
@@ -1,3 +1,3 @@
 const isAbsent = (value: any): value is undefined | null => value == null;
 
-export default isAbsent
+export default isAbsent;
