+++ src/util/isSchema.ts
@@ -1,4 +1,4 @@
-import type { ISchema } from './types';
+import type { ISchema } from '../types';
 
 const isSchema = (obj: any): obj is ISchema<any> => obj && obj.__isYupSchema__;
