+++ tsconfig.json
@@ -1,9 +1,8 @@
 {
   "extends": "@4c/tsconfig/web",
   "compilerOptions": {
-    "allowJs": true,
     "rootDir": "src",
     "strictFunctionTypes": true
   },
-  "include": ["src"]
+  "include": ["src/**/*.ts"]
 }
